requirements.txt
# Core dependencies
openai>=2.9.0
chromadb>=1.3.0
tiktoken>=0.12.0

# Web API (FastAPI backend for React UI)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
python-multipart>=0.0.6

# HTML parsing
beautifulsoup4>=4.14.0

# Testing
pytest>=9.0.0
pytest-cov>=7.0.0

# Config
python-dotenv>=1.2.0
PyYAML>=6.0.0
